requirements.txt
streamlit==1.31.1
pytesseract==0.3.10
pillow==10.2.0
clipboard==0.0.4
easyocr==1.7.1
numpy==1.26.2
